services/audio-processing/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Audio processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# Image processing (for waveforms)
Pillow==10.1.0

# Storage
aioboto3==12.1.0  # For S3/MinIO support

# Authentication
pyjwt==2.8.0

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0